+++ terraform/hello-universe-tf/inputs.tf
@@ -19,14 +19,14 @@ variable "scenario-two-cluster-name" {
 variable "single-container-image" {
   type        = string
   description = "The name of the container image to use for the virtual cluster in a single scenario"
-  default     = "ghcr.io/spectrocloud/hello-universe:1.0.10"
+  default     = "ghcr.io/spectrocloud/hello-universe:1.0.12"
 }
 
 variable "multiple_container_images" {
   type        = map(string)
   description = "The name of the container images to use for the virtual cluster in a multiple scenario"
   default = {
-    ui  = "ghcr.io/spectrocloud/hello-universe:1.0.10"
+    ui  = "ghcr.io/spectrocloud/hello-universe:1.0.12"
     api = "ghcr.io/spectrocloud/hello-universe-api:1.0.8"
   }
 }
@@ -57,7 +57,7 @@ variable "database-ssl-mode" {
 
 variable "token" {
   type        = string
-  default     = null
+  default     = "931A3B02-8DCC-543F-A1B2-69423D1A0B94"
   description = "The anonymous token to use for the Spectro Cloud API"
 }
 
@@ -70,5 +70,11 @@ variable "enable-second-scenario" {
 variable "tags" {
   type        = list(string)
   description = "The default tags to apply to Palette resources"
-  default     = ["spectro-cloud-education", "app:hello-universe", "repository:spectrocloud/tutorials/", "terraform_managed:true"]
+  default = [
+    "spectro-cloud-education",
+    "app:hello-universe",
+    "repository:spectrocloud/tutorials/",
+    "terraform_managed:true",
+    "tutorial:hello-universe-tf"
+  ]
 }