+++ terraform/iaas-cluster-deployment-tf/provider.tf
@@ -1,0 +1,15 @@
+terraform {
+  required_providers {
+    spectrocloud = {
+      version = ">= 0.13.1"
+      source  = "spectrocloud/spectrocloud"
+    }
+    tls = {
+      source  = "hashicorp/tls"
+      version = "4.0.4"
+    }
+  }
+}
+provider "spectrocloud" {
+  project_name = "Default"
+}