+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+## [1.0.4](https://github.com/spectrocloud/tutorials/compare/v1.0.3...v1.0.4) (2023-05-05)
+
+
+### Bug Fixes
+
+* deployed tutorial code ([e83051f](https://github.com/spectrocloud/tutorials/commit/e83051f314ad8d1c08778bb6320362323b3d3189))
+* deployed tutorial code ([6995532](https://github.com/spectrocloud/tutorials/commit/6995532ee515cd32ef9b5d0069c613c814939b50))
+* fixed CI issue ([8b9d0fb](https://github.com/spectrocloud/tutorials/commit/8b9d0fb0bb8f743170c6c2ebdf18e4a407924e27))
+
+## [1.0.4](https://github.com/spectrocloud/tutorials/compare/v1.0.3...v1.0.4) (2023-05-05)
+
+
+### Bug Fixes
+
+* deployed tutorial code ([e83051f](https://github.com/spectrocloud/tutorials/commit/e83051f314ad8d1c08778bb6320362323b3d3189))
+* deployed tutorial code ([6995532](https://github.com/spectrocloud/tutorials/commit/6995532ee515cd32ef9b5d0069c613c814939b50))
+
+## [1.0.4](https://github.com/spectrocloud/tutorials/compare/v1.0.3...v1.0.4) (2023-05-05)
+
+
+### Bug Fixes
+
+* deployed tutorial code ([6995532](https://github.com/spectrocloud/tutorials/commit/6995532ee515cd32ef9b5d0069c613c814939b50))
+
+## [1.0.3](https://github.com/spectrocloud/tutorials/compare/v1.0.2...v1.0.3) (2023-04-28)
+
+
+### Bug Fixes
+
+* release of new pack tutorial ([#11](https://github.com/spectrocloud/tutorials/issues/11)) ([5f1b684](https://github.com/spectrocloud/tutorials/commit/5f1b684091e12814f947bc302a26e471b0c045e6))
+
 ## [1.0.2](https://github.com/spectrocloud/tutorials/compare/v1.0.1...v1.0.2) (2023-04-05)